renamed je form steps
