+++ components/journal/journal-entry-form/FormStepProgress.tsx
@@ -87,7 +87,7 @@ export function FormStepProgress({
                   {count}
                 </Badge>
               )}
-              {step.type === "reward" && (
+              {step.type === "bonus" && (
                 <Badge
                   variant="outline"
                   className={`${bgColor} absolute -top-1 -right-1 text-[0.6rem] px-1 py-0 min-w-[1.2rem] h-[1.2rem] flex items-center justify-center text-white`}
