+++ components/journal/journal-entry-form/form-steps/steps/Highlights.tsx
@@ -1,11 +1,11 @@
 import { useFormContext } from "react-hook-form";
 import { FormStepTemplate } from "@components/journal/journal-entry-form/form-steps/FormStepTemplate";
 import { TextAreaList } from "@components/ui/textarea-list";
-import { StepScoreDisplay } from "./StepScoreDisplay";
+import { StepScoreDisplay } from "../StepScoreDisplay";
 // import { calculateStepScore } from "@lib/score";
 import type { JournalEntry } from "@models/types";
 
-export function DailyHighlights() {
+export function Highlights() {
   const { watch, setValue } = useFormContext<JournalEntry>();
 
   const dailyHighlights = watch("nightEntry.dailyHighlights");
